+++ scripts/configure.js
@@ -152,7 +152,7 @@ Inquirer.prompt([
 
     // Fetch configuration from the panel
     console.log('Fetching configuration from panel.');
-    Request.get(`${params.panelurl.value}/remote/configuration/${params.token.value}`, (error, response, body) => {
+    Request.get(`${params.panelurl.value}/daemon/configure/${params.token.value}`, (error, response, body) => {
         if (!error) {
             // response should always be JSON
             const jsonBody = JSON.parse(body);